+++ src/shared/models/post/create-post.dto.ts
@@ -1,7 +1,5 @@
-import { ID } from 'scalars';
-
 export class CreatePostDto {
   public title?: string;
   public content: string;
-  public creator: ID;
+  public creator: string;
 }