+++ src/shared/models/user/update-user.dto.ts
@@ -1,5 +1,3 @@
-import { ID } from 'scalars';
-
 export class UpdateUserDto {
   name: string;
 }