+++ src/shared/models/user/user-with-password.model.ts
@@ -1,7 +1,7 @@
 import { UserModel } from './user.model';
 
 export class UserWithPasswordModel extends UserModel {
-  static parseToUser(userWithPassword: UserWithPasswordModel): UserModel {
+  static parseToUser(userWithPassword: UserWithPasswordModel) {
     const user: UserModel = {
       id: userWithPassword.id,
       name: userWithPassword.name,