+++ src/shared/typings/scalars.d.ts
@@ -1,3 +1,0 @@
-declare module 'scalars' {
-  export type ID = string;
-}